requirements.txt
matplotlib
pycocotools
Pillow
numpy
tqdm
opencv-python
requests
